services/generator/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
pyyaml==6.0.2
google-cloud-pubsub==2.21.5
jsonschema==4.22.0
python-dateutil==2.9.0.post0
